+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDYZDQ5P99W59X7DPXS1PX25"
+        "01KE6FM1C68YGJYT0TA31JASEK"
     )
     r"""The unique identifier of the data source"""