+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -152,7 +152,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdyzdpqck7zeq8mzdf49984s"
+        "conv_01ke6fm0yjeg4khjfccydc5yad"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """