+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDYZDQ5NXWEHGE7SVXR26G9B"
+        "01KE6FM1C5YNBS6GCKRQH5TYNV"
     )
     r"""The unique identifier of the data source"""