+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -1463,7 +1463,7 @@ class AgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KDYZDPN3NTSG5QD11QSXM0PA"
+    id: Optional[str] = "01KE6FM0WCYMG0HD1W5VFYMK3Q"
 
     description: Optional[str] = None