+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -150,7 +150,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdyzdpqg3a39kqtgskpzyv9s"
+        "conv_01ke6fm0ypz5v7kj9vp6ehac5e"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """