+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDYZDQ5SZ2RHKCK0673RXCR4"
+        "01KE6FM1CAE03SZX2BB25QH2C3"
     )
     r"""The unique identifier of the data source"""